data type issue

+++ src/getangle.c
@@ -18,7 +18,7 @@
 
 int main(){
 
-  double aX,aY,aZ,pitch;
+  int aX,aY,aZ,pitch;
 
 // setup i2c
   int devAccel = wiringPiI2CSetup(0x53);
@@ -50,8 +50,8 @@ int main(){
      aY = aY * 0.00390625;
 
      pitch = (atan2(aX,sqrt(aY*aY+aZ*aZ)) * 180.0) / PI;
-     printf("%lf,%lf,%lf\n",aX,aY,aZ );
-     printf("pitch = %lf\n",pitch);
+     printf("%d,%d,%d\n",aX,aY,aZ );
+     printf("pitch = %d\n",pitch);
    }
   return(0);
 }
